"Finished" duplicate updater
I'm just truncating all $id values with a length > 6 because there's a
weird bug where MySQL thinks that an ID of 012682 and 01268223948 are
the same thing. (Note that $id is of type string, too)

Adding extra characters to a string isn't a problem if you can't do it,
right?

TODO: **Make sure** to note the length of the ID given in the return
HTML page to the client.
TODO: Add a system to add leading zeroes to an ID that's too short to
make its length = 6.